+++ src/Registry/Source/JsonFile.php
@@ -24,14 +24,16 @@
  *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Router
  */
+
 namespace CeusMedia\Router\Registry\Source;
 
+use CeusMedia\Common\FS\File\JSON\Reader as JsonFileReader;
+use CeusMedia\Common\FS\File\JSON\Writer as JsonFileWriter;
 use CeusMedia\Router\Registry;
-use CeusMedia\Router\Registry\Source\AbstractSource;
-use CeusMedia\Router\Registry\Source\SourceInterface;
 use CeusMedia\Router\Route;
 use CeusMedia\Router\Route\Factory as RouteFactory;
-use FS_File_JSON_Reader as JsonFileReader;
+use OutOfRangeException;
+//use RuntimeException;
 
 /**
  *	...
@@ -49,13 +51,13 @@ class JsonFile extends AbstractSource implements SourceInterface
 	{
 		if( !file_exists( $this->resource ) )
 			return -1;
-//			throw new \RuntimeException( 'JSON file "'.$this->resource.'" is not existing' );
+//			throw new RuntimeException( 'JSON file "'.$this->resource.'" is not existing' );
 		$counter	= 0;
 		$data		= JsonFileReader::load( $this->resource, FALSE );
 		$factory	= new RouteFactory();
 		foreach( (array) $data as $item ){
 			if( !isset( $item->pattern ) )
-				throw new \OutOfRangeException( 'Route set is missing pattern' );
+				throw new OutOfRangeException( 'Route set is missing pattern' );
 			$options	= array(
 				'method'		=> isset( $item->method ) ? $item->method : NULL,
 				'controller'	=> isset( $item->controller ) ? $item->controller : NULL,
@@ -65,6 +67,8 @@ class JsonFile extends AbstractSource implements SourceInterface
 				$options['mode']	= Route::getModeFromKey( $item->mode );
 			if( isset( $item->roles ) && strlen( trim( $item->roles ) ) > 0 )
 				$options['roles']	= preg_split( "/, */", trim( $item->roles ) );
+			if( isset( $item->priority ) && strlen( trim( $item->priority ) ) > 0 )
+				$options['priority']	= Route::getPriorityFromKey( $item->priority );
 			$registry->add( $factory->create( $item->pattern, $options ) );
 			$counter++;
 		}
@@ -73,18 +77,20 @@ class JsonFile extends AbstractSource implements SourceInterface
 
 	public function save( Registry $registry ): int
 	{
-		$data	= array();
+		$data	= [];
 		foreach( $registry->index() as $route ){
 			$mode	= Route::getModeKey( $route->getMode() );
-			$item	= array();
+			$item	= [];
 			if( strlen( $mode ) > 0 )
 				$item['mode']	= $mode;
 			$item['controller']	= $route->getController();
 			$item['action']		= $route->getAction();
 			$item['pattern']	= $route->getPattern();
 			$item['method']		= $route->getMethod();
+			if( $route->getPriority() !== Route::PRIORITY_NORMAL )
+				$item['priority']	= Route::getPriorityKey( $route->getPriority() );
 			$data[]	= $item;
 		}
-		return \FS_File_JSON_Writer::save( $this->resource, $data, TRUE );
+		return JsonFileWriter::save( $this->resource, $data, TRUE );
 	}
 }