+++ src/Route.php
@@ -24,7 +24,12 @@
  *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Router
  */
+
 namespace CeusMedia\Router;
+
+use DomainException;
+use InvalidArgumentException;
+use RangeException;
 
 /**
  *	...
@@ -32,7 +37,7 @@ namespace CeusMedia\Router;
  *	@category		Library
  *	@package		CeusMedia_Router
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
- *	@copyright		2016-2020 Christian Würker
+ *	@copyright		2016-2022 Christian Würker
  *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Router
  */
@@ -48,49 +53,80 @@ class Route
 	const MODE_KEY_EVENT		= 'event';
 	const MODE_KEY_FORWARD		= 'forward';
 
-	const MODES_BY_KEYS			= array(
-		'unknown'				=> self::MODE_UNKNOWN,
-		'controller'			=> self::MODE_CONTROLLER,
-		'event'					=> self::MODE_EVENT,
-		'forward'				=> self::MODE_FORWARD,
-	);
+	const MODES_BY_KEYS			= [
+		self::MODE_KEY_UNKNOWN		=> self::MODE_UNKNOWN,
+		self::MODE_KEY_CONTROLLER	=> self::MODE_CONTROLLER,
+		self::MODE_KEY_EVENT		=> self::MODE_EVENT,
+		self::MODE_KEY_FORWARD		=> self::MODE_FORWARD,
+	];
 
-	const MODE_KEYS				= array(
-		self::MODE_UNKNOWN		=> 'unknown',
-		self::MODE_CONTROLLER	=> 'controller',
-		self::MODE_EVENT		=> 'event',
-		self::MODE_FORWARD		=> 'forward',
-	);
+	const MODE_KEYS				= [
+		self::MODE_UNKNOWN		=> self::MODE_KEY_UNKNOWN,
+		self::MODE_CONTROLLER	=> self::MODE_KEY_CONTROLLER,
+		self::MODE_EVENT		=> self::MODE_KEY_EVENT,
+		self::MODE_FORWARD		=> self::MODE_KEY_FORWARD,
+	];
 
-	/** @var	string			$method				Request method of route */
-	protected $method			= 'GET';
+	const PRIORITY_EARLIEST		= 1;
+	const PRIORITY_EARLIER		= 2;
+	const PRIORITY_NORMAL		= 3;
+	const PRIORITY_LATER		= 4;
+	const PRIORITY_LATEST		= 5;
 
-	/** @var	integer			$mode				Mode of route: 1:controller, 2:event: 3: forward */
-	protected $mode				= self::MODE_UNKNOWN;
+	const PRIORITY_KEY_EARLIEST		= 'earliest';
+	const PRIORITY_KEY_EARLIER		= 'earlier';
+	const PRIORITY_KEY_NORMAL		= 'normal';
+	const PRIORITY_KEY_LATER		= 'later';
+	const PRIORITY_KEY_LATEST		= 'latest';
 
-	/** @var	string			$pattern			... */
-	protected $pattern			= '';
+	const PRIORITIES_BY_KEYS			= [
+		self::PRIORITY_KEY_EARLIEST		=> self::PRIORITY_EARLIEST,
+		self::PRIORITY_KEY_EARLIER		=> self::PRIORITY_EARLIER,
+		self::PRIORITY_KEY_NORMAL		=> self::PRIORITY_NORMAL,
+		self::PRIORITY_KEY_LATER		=> self::PRIORITY_LATER,
+		self::PRIORITY_KEY_LATEST		=> self::PRIORITY_LATEST,
+	];
 
-	/** @var	string			$controller			Controller class attached to route */
-	protected $controller		= '';
+	const PRIORITY_KEYS			= [
+		self::PRIORITY_EARLIEST		=> self::PRIORITY_KEY_EARLIEST,
+		self::PRIORITY_EARLIER		=> self::PRIORITY_KEY_EARLIER,
+		self::PRIORITY_NORMAL		=> self::PRIORITY_KEY_NORMAL,
+		self::PRIORITY_LATER		=> self::PRIORITY_KEY_LATER,
+		self::PRIORITY_LATEST		=> self::PRIORITY_KEY_LATEST,
+	];
 
-	/** @var	string			$action				Action if controller attached to route */
-	protected $action			= '';
+	/** @var	string				$method				Request method of route */
+	protected string $method		= 'GET';
 
-	/** @var	array			$arguments			Key in cache */
-	protected $arguments		= array();
+	/** @var	integer				$mode				Mode of route: 1:controller, 2:event: 3: forward */
+	protected int $mode				= self::MODE_UNKNOWN;
 
-	/** @var	array			$roles				List of role keys to limit route access to, empty means no limits */
-	protected $roles			= array();
+	/** @var	string				$pattern			... */
+	protected string $pattern		= '';
 
-	/** @var	?Route			$origin				... */
-	protected $origin;
+	/** @var	string				$controller			Controller class attached to route */
+	protected string $controller	= '';
 
-	/** @var	string			$target				Target address for route of type "forward" */
-	protected $target;
+	/** @var	string				$action				Action if controller attached to route */
+	protected string $action		= '';
 
-	/** @var	array			$supportedMethods	Allowed request methods */
-	protected $supportedMethods	= array(
+	/** @var	array				$arguments			Key in cache */
+	protected array $arguments		= [];
+
+	/** @var	array				$roles				List of role keys to limit route access to, empty means no limits */
+	protected array $roles			= [];
+
+	/** @var	?Route				$origin				... */
+	protected ?Route $origin		= NULL;
+
+	/** @var	int					$priority			Resolver priority: 1:earliest, 2:earlier, 3:normal, 4:later, 5:latest */
+	protected int $priority			= self::PRIORITY_NORMAL;
+
+	/** @var	string				$target				Target address for route of type "forward" */
+	protected string $target		= '';
+
+	/** @var	array				$supportedMethods	Allowed request methods */
+	public array $supportedMethods		= [
 		'CLI',
 		'GET',
 		'HEAD',
@@ -98,7 +134,46 @@ class Route
 		'PUT',
 		'DELETE',
 		'OPTIONS',
-	);
+	];
+
+	public static function getModeFromKey( string $mode, bool $strict = TRUE ): int
+	{
+		$mode	= strtolower( $mode );
+		if( array_key_exists( $mode, self::MODES_BY_KEYS ) )
+			return self::MODES_BY_KEYS[$mode];
+		if( $strict )
+			throw new RangeException( 'Invalid mode key: '.$mode );
+		return self::MODE_UNKNOWN;
+	}
+
+	public static function getModeKey( int $mode, bool $strict = TRUE ): string
+	{
+		if( array_key_exists( $mode, self::MODE_KEYS) )
+			return self::MODE_KEYS[$mode];
+		if( $strict )
+			throw new RangeException( 'Invalid mode: '.$mode );
+		return self::MODE_KEY_UNKNOWN;
+	}
+
+	public static function getPriorityFromKey( string $priority ): int
+	{
+		$mode	= strtolower( $priority );
+		if( array_key_exists( $priority, self::PRIORITIES_BY_KEYS ) )
+			return self::PRIORITIES_BY_KEYS[$priority];
+		throw new RangeException( 'Invalid priority key: '.$priority );
+	}
+
+	/**
+	 *	@param		int			$priority
+	 *	@return		string
+	 *	@throws		RangeException		if priority is not supported (= within [1-5])
+	 */
+	public static function getPriorityKey( int $priority ): string
+	{
+		if( array_key_exists( $priority, self::PRIORITY_KEYS ) )
+			return self::PRIORITY_KEYS[$priority];
+		throw new RangeException( 'Invalid priority: '.$priority );
+	}
 
 	public function __construct( string $pattern, string $method = NULL, int $mode = NULL )
 	{
@@ -139,25 +214,6 @@ class Route
 		return $this->mode;
 	}
 
-	public static function getModeFromKey( string $mode, bool $strict = TRUE ): int
-	{
-		$mode	= strtolower( $mode );
-		if( array_key_exists( $mode, self::MODES_BY_KEYS ) )
-			return self::MODES_BY_KEYS[$mode];
-		if( $strict )
-			throw new \RangeException( 'Invalid mode key: '.$mode );
-		return self::MODE_UNKNOWN;
-	}
-
-	public static function getModeKey( int $mode, bool $strict = TRUE ): string
-	{
-		if( array_key_exists( $mode, self::MODE_KEYS) )
-			return self::MODE_KEYS[$mode];
-		if( $strict )
-			throw new \RangeException( 'Invalid mode: '.$mode );
-		return self::MODE_KEY_UNKNOWN;
-	}
-
 	public function getOrigin(): ?Route
 	{
 		return $this->origin;
@@ -166,6 +222,11 @@ class Route
 	public function getPattern(): string
 	{
 		return $this->pattern;
+	}
+
+	public function getPriority(): int
+	{
+		return $this->priority;
 	}
 
 	public function getRoles(): array
@@ -200,7 +261,7 @@ class Route
 	public function setAction( string $action ): self
 	{
 		if( preg_match( '/^(_|[a-z0-9])+$/i', $action ) === 0 )
-			throw new \InvalidArgumentException( 'Action must be a valid method name' );
+			throw new InvalidArgumentException( 'Action must be a valid method name' );
 		$this->action		= $action;
 		return $this;
 	}
@@ -226,7 +287,7 @@ class Route
 	public function setController( string $controller ): self
 	{
 		if( preg_match( '/^(_|\\\|[a-z0-9])+$/i', $controller ) === 0 )
-			throw new \InvalidArgumentException( 'Controller must be a valid class name' );
+			throw new InvalidArgumentException( 'Controller must be a valid class name' );
 		$this->controller	= $controller;
 		return $this;
 	}
@@ -236,11 +297,11 @@ class Route
 	 *	@access		public
 	 *	@param		string		$method			...
 	 *	@return		Route
-	 *	@throws		\DomainException			if given method is invalid or not supported
+	 *	@throws		DomainException			if given method is invalid or not supported
 	 */
 	public function setMethod( string $method ): self
 	{
-		$validMethods	= array();
+		$validMethods	= [];
 		$methods		= preg_split( '/\s*(,|\|)\s*/', strtoupper( trim( $method ) ) );
 		if( $methods !== FALSE ){
 			if( in_array( '*', $methods, TRUE ) )
@@ -250,7 +311,7 @@ class Route
 					if( strlen( trim( $item ) ) === 0 )
 						continue;
 					if( !in_array( $item, $this->supportedMethods, TRUE ) )
-						throw new \DomainException( 'Invalid method: '.$item );
+						throw new DomainException( 'Invalid method: '.$item );
 					$validMethods[]	= $item;
 				}
 			}
@@ -264,12 +325,11 @@ class Route
 	 *	@access		public
 	 *	@param		int			$mode			Mode as constant value (int)
 	 *	@return		Route
-	 *	@throws		\DomainException			if given mode value is no a valid constant value (int)
+	 *	@throws		DomainException			if given mode value is no a valid constant value (int)
 	 */
 	public function setMode( int $mode ): self
 	{
-		if( preg_match( '/^[0-9]+$/', (string) $mode ) === 0 )
-			throw new \DomainException( 'Invalid mode: '.$mode );
+		self::getModeKey( $mode, TRUE );
 		$this->mode	= $mode;
 		return $this;
 	}
@@ -295,8 +355,22 @@ class Route
 	public function setPattern( string $pattern ): self
 	{
 		if( $this->method !== 'CLI' && preg_match( '/\s/', $pattern ) > 0 )
-			throw new \InvalidArgumentException( 'Route pattern must not contain whitespace ('.$pattern.')' );
+			throw new InvalidArgumentException( 'Route pattern must not contain whitespace ('.$pattern.')' );
 		$this->pattern		= $pattern;
+		return $this;
+	}
+
+	/**
+	 *	...
+	 *	@access		public
+	 *	@param		int			$priority
+	 *	@return		Route
+	 *	@throws		RangeException		if priority is not supported (= within [1-5])
+	 */
+	public function setPriority( int $priority ): self
+	{
+		self::getPriorityKey( $priority );
+		$this->priority		= $priority;
 		return $this;
 	}
 
@@ -331,7 +405,7 @@ class Route
 	 */
 	public function toArray(): array
 	{
-		return array(
+		return [
 			'id'			=> $this->getId(),
 			'mode'			=> $this->mode,
 			'method'		=> $this->method,
@@ -342,6 +416,7 @@ class Route
 			'roles'			=> $this->roles,
 			'origin'		=> $this->origin,
 			'target'		=> $this->target,
-		);
+			'priority'		=> $this->priority,
+		];
 	}
 }