+++ src/Route/Factory.php
@@ -24,6 +24,7 @@
  *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Router
  */
+
 namespace CeusMedia\Router\Route;
 
 use CeusMedia\Router\Route;
@@ -40,21 +41,21 @@ use CeusMedia\Router\Route;
  */
 class Factory
 {
-	/** @var	?int		$defaultMode	Mode to set by default for new route */
-	protected $defaultMode;
+	/** @var	int|NULL				$defaultMode	Mode to set by default for new route */
+	protected ?int $defaultMode			= NULL;
 
-	/** @var	?string		$defaultMethod	Request method  to set by default for new route */
-	protected $defaultMethod;
+	/** @var	string|NULL				$defaultMethod	Request method  to set by default for new route */
+	protected ?string $defaultMethod	= NULL;
 
-	public function create( string $pattern, array $options = array() ): Route
+	public function create( string $pattern, array $options = [] ): Route
 	{
-		$options	= array_merge( array(
+		$options	= array_merge( [
 			'method'		=> $this->defaultMethod,
 			'mode'			=> $this->defaultMode,
 			'controller'	=> NULL,
 			'action'		=> NULL,
-			'roles'			=> array(),
-		), $options );
+			'roles'			=> [],
+		], $options );
 
 		$route	= new Route( $pattern, $options['method'], $options['mode'] );
 		if( isset( $options['controller'] ) && strlen( trim( $options['controller'] ) ) > 0 )
@@ -69,6 +70,11 @@ class Factory
 					if( preg_split( '/\s*,\s*/', $options['roles'] ) !== FALSE )
 						$route->setRoles( preg_split( '/\s*,\s*/', $options['roles'] ) );
 			}
+		}
+		if( isset( $options['priority'] ) && strlen( trim( $options['priority'] ) ) > 0 ){
+			if( !is_int( $options['priority'] ) )
+				$options['priority']	= Route::getPriorityFromKey( $options['priority'] );
+			$route->setPriority( $options['priority'] );
 		}
 		return $route;
 	}